Refactor MCP algorithm discovery to use Proxy plugin catalogs
Remove compatibility fallback catalogs and WorkflowQueryResult.
Validate algorithms against the current Proxy and clarify recommendation
provenance and capability risks.
